Working Migration from B5 -> B6.  New variables are done!
Former-commit-id: 7cc44af006d197f77c4a7611b2afb0bff2145f5b

+++ Intersect (Core)/GameObjects/Events/VariableMod.cs
@@ -16,8 +16,8 @@ namespace Intersect.GameObjects.Events
     public class IntegerVariableMod : VariableMod
     {
         public VariableMods ModType { get; set; } = VariableMods.Set;
-        public int Value { get; set; }
-        public int HighValue { get; set; }
+        public long Value { get; set; }
+        public long HighValue { get; set; }
         public Guid DupVariableId { get; set; }
     }
